requirements.txt
# Core MCP dependencies
fastmcp>=0.4.0
mcp[cli]>=1.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# File format support
openpyxl>=3.1.0      # Excel files
pyarrow>=14.0.0      # Parquet files
xlrd>=2.0.0          # Legacy Excel files
odfpy>=1.4.0         # OpenDocument spreadsheets
lxml>=4.9.0          # XML/HTML parsing
html5lib>=1.1        # HTML parsing
tables>=3.8.0        # HDF5 support (optional, comment out if issues)

# Server dependencies for SSE
uvicorn[standard]>=0.25.0
httpx>=0.25.0
aiofiles>=23.0.0
python-multipart>=0.0.6
sse-starlette>=1.8.0

# Visualization (optional but recommended)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0       # For interactive plots (optional)
kaleido>=0.2.1       # For static image export from plotly (optional)

# Text-based visualization
tabulate>=0.9.0      # Pretty text tables
rich>=13.0.0         # Rich text formatting
asciichartpy>=1.5.0  # ASCII charts

# Environment and configuration
python-dotenv>=1.0.0

# Utilities
aiohttp>=3.9.0       # Async HTTP client
nest-asyncio>=1.5.0  # Nested async loops
psutil>=5.9.0        # System monitoring
chardet==5.2.0       # Character encoding detection

# Development dependencies (optional, for development only)
# Uncomment these if you want to include dev dependencies in main requirements
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.7.0
# ipython>=8.0.0
# jupyter>=1.0.0
